// src/app/shared/interfaces/relatedDataServiceAndProduct.interface.ts
export interface CategoryType {
  categoryTypeId: number;
  name: string;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date;
}

export interface StateType {
  stateTypeId: number;
  name: string;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date;
}

export interface BedType {
  bedTypeId: number;
  name: string;
  createdAt: Date;
  updatedAt: Date;
  deletedAt: Date;
}